libglusterfs: generic buffer infrastructure
Generic buffer is the simple concept to implement basic operations
on the memory area of the known size. Basically the buffer is just
a pair of raw C pointer and size; if the last byte of the pointed
area is '\0', this area can be treated as a C string.

Generic buffers are intended to be a function- or block-scoped, and
provides simple API mostly covers initialization, copying, clearing,
concatenation and basic formatted output.

Signed-off-by: Dmitry Antipov <dantipov@cloudlinux.com>
Updates: #1000